feat: enable low-risk linters (unconvert, wastedassign, usestdlibvars)
Phase 1 of linting improvements - enable linters that catch real bugs
and improve code quality with minimal false positives.

Changes:
- Enable unconvert: Remove unnecessary type conversions (123 fixes)
- Enable wastedassign: Fix unused variable assignments (3 fixes)
- Enable usestdlibvars: Use stdlib constants (98 fixes)
  - HTTP status codes: use http.StatusOK, http.StatusNotFound, etc.
  - HTTP methods: use http.MethodGet, http.MethodPost, etc.
  - Time layouts: use time.DateTime, time.DateOnly, time.TimeOnly

Total: 224 linting issues resolved across 72 files

All tests pass. No regressions introduced.

+++ internal/adapters/v0_0_40/job_adapter.go
@@ -242,7 +242,7 @@ func (a *JobAdapter) Submit(ctx context.Context, job *types.JobCreate) (*types.J
 
 	// Node count
 	if job.Nodes > 0 {
-		nodes := int32(job.Nodes)
+		nodes := job.Nodes
 		jobDesc.MinimumNodes = &nodes
 		jobDesc.MaximumNodes = &nodes
 	}
@@ -303,7 +303,7 @@ func (a *JobAdapter) Submit(ctx context.Context, job *types.JobCreate) (*types.J
 
 	// Extract job ID
 	if resp.JSON200.JobId != nil {
-		submitResp.JobID = int32(*resp.JSON200.JobId)
+		submitResp.JobID = *resp.JSON200.JobId
 	}
 
 	// Extract warnings if any
@@ -454,7 +454,7 @@ func (a *JobAdapter) Hold(ctx context.Context, req *types.JobHoldRequest) error 
 	update := &types.JobUpdate{
 		Priority: func() *int32 {
 			if req.Priority != 0 {
-				priority := int32(req.Priority)
+				priority := req.Priority
 				return &priority
 			}
 			return nil
